move workspace/desktop from WSSwitcher=>Pager
